+++ Assets/Scripts/ConnectionManagement/ConnectionState/ClientConnectingState.cs
@@ -1,9 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using Unity.BossRoom.UnityServices.Lobbies;
-using Unity.Multiplayer.Samples.BossRoom;
 using Unity.Multiplayer.Samples.Utilities;
-using Unity.Netcode.Transports.UTP;
 using UnityEngine;
 using VContainer;
 
@@ -20,6 +18,13 @@ namespace Unity.BossRoom.ConnectionManagement
         protected LobbyServiceFacade m_LobbyServiceFacade;
         [Inject]
         protected LocalLobby m_LocalLobby;
+        ConnectionMethodBase m_ConnectionMethod;
+
+        public ClientConnectingState Configure(ConnectionMethodBase baseConnectionMethod)
+        {
+            m_ConnectionMethod = baseConnectionMethod;
+            return this;
+        }
 
         public override void Enter()
         {
@@ -38,6 +43,12 @@ namespace Unity.BossRoom.ConnectionManagement
 
         public override void OnClientDisconnect(ulong _)
         {
+            // client ID is for sure ours here
+            StartingClientFailedAsync();
+        }
+
+        protected void StartingClientFailedAsync()
+        {
             m_ConnectStatusPublisher.Publish(ConnectStatus.StartClientFailed);
             m_ConnectionManager.ChangeState(m_ConnectionManager.m_Offline);
         }
@@ -48,52 +59,30 @@ namespace Unity.BossRoom.ConnectionManagement
             m_ConnectionManager.ChangeState(m_ConnectionManager.m_DisconnectingWithReason);
         }
 
-        protected async Task ConnectClientAsync()
+
+        internal async Task ConnectClientAsync()
         {
-            bool success = true;
-            if (m_LobbyServiceFacade.CurrentUnityLobby != null)
+            try
             {
-                success = await JoinRelayServerAsync();
-            }
+                // Setup NGO with current connection method
+                await m_ConnectionMethod.SetupClientConnectionAsync();
 
-            if (success)
-            {
-                success = m_ConnectionManager.NetworkManager.StartClient();
-            }
+                // NGO's StartClient launches everything
+                if (!m_ConnectionManager.NetworkManager.StartClient())
+                {
+                    throw new Exception("NetworkManager StartClient failed");
+                }
 
-            if (success)
-            {
                 SceneLoaderWrapper.Instance.AddOnSceneEventCallback();
                 m_ConnectionManager.RegisterCustomMessages();
             }
-            else
-            {
-                OnClientDisconnect(0);
-            }
-        }
-
-        async Task<bool> JoinRelayServerAsync()
-        {
-            Debug.Log($"Setting Unity Relay client with join code {m_LocalLobby.RelayJoinCode}");
-
-            try
-            {
-                var (ipv4Address, port, allocationIdBytes, allocationId, connectionData, hostConnectionData, key) =
-                    await UnityRelayUtilities.JoinRelayServerFromJoinCode(m_LocalLobby.RelayJoinCode);
-
-                await m_LobbyServiceFacade.UpdatePlayerRelayInfoAsync(allocationId.ToString(), m_LocalLobby.RelayJoinCode);
-                var utp = (UnityTransport)m_ConnectionManager.NetworkManager.NetworkConfig.NetworkTransport;
-                utp.SetClientRelayData(ipv4Address, port, allocationIdBytes, key, connectionData, hostConnectionData, isSecure: true);
-            }
             catch (Exception e)
             {
-                Debug.Log($"Relay join failed: {e.Message}");
-                //leave the lobby if relay failed for some reason
-                await m_LobbyServiceFacade.EndTracking();
-                return false;
+                Debug.LogError("Error connecting client, see following exception");
+                Debug.LogException(e);
+                StartingClientFailedAsync();
+                throw;
             }
-
-            return true;
         }
     }
 }