+++ Assets/Scripts/ConnectionManagement/ConnectionState/OfflineState.cs
@@ -3,9 +3,6 @@ using Unity.BossRoom.ConnectionManagement;
 using Unity.BossRoom.UnityServices.Lobbies;
 using Unity.BossRoom.Utils;
 using Unity.Multiplayer.Samples.Utilities;
-using Unity.Netcode.Transports.UTP;
-using Unity.Services.Authentication;
-using Unity.Services.Core;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using VContainer;
@@ -22,6 +19,8 @@ namespace UUnity.BossRoom.ConnectionManagement
         LobbyServiceFacade m_LobbyServiceFacade;
         [Inject]
         ProfileManager m_ProfileManager;
+        [Inject]
+        LocalLobby m_LocalLobby;
 
         const string k_MainMenuSceneName = "MainMenu";
 
@@ -39,55 +38,28 @@ namespace UUnity.BossRoom.ConnectionManagement
 
         public override void StartClientIP(string playerName, string ipaddress, int port)
         {
-            var utp = (UnityTransport)m_ConnectionManager.NetworkManager.NetworkConfig.NetworkTransport;
-            utp.SetConnectionData(ipaddress, (ushort)port);
-            SetConnectionPayload(GetPlayerId(), playerName);
-            m_ConnectionManager.ChangeState(m_ConnectionManager.m_ClientConnecting);
+            var connectionMethod = new ConnectionMethodIP(ipaddress, (ushort)port, m_ConnectionManager, m_ProfileManager, playerName);
+            m_ConnectionManager.m_ClientReconnecting.Configure(connectionMethod);
+            m_ConnectionManager.ChangeState(m_ConnectionManager.m_ClientConnecting.Configure(connectionMethod));
         }
 
         public override void StartClientLobby(string playerName)
         {
-            SetConnectionPayload(GetPlayerId(), playerName);
-            m_ConnectionManager.ChangeState(m_ConnectionManager.m_ClientConnecting);
+            var connectionMethod = new ConnectionMethodRelay(m_LobbyServiceFacade, m_LocalLobby, m_ConnectionManager, m_ProfileManager, playerName);
+            m_ConnectionManager.m_ClientReconnecting.Configure(connectionMethod);
+            m_ConnectionManager.ChangeState(m_ConnectionManager.m_ClientConnecting.Configure(connectionMethod));
         }
 
         public override void StartHostIP(string playerName, string ipaddress, int port)
         {
-            var utp = (UnityTransport)m_ConnectionManager.NetworkManager.NetworkConfig.NetworkTransport;
-            utp.SetConnectionData(ipaddress, (ushort)port);
-
-            SetConnectionPayload(GetPlayerId(), playerName);
-            m_ConnectionManager.ChangeState(m_ConnectionManager.m_StartingHost);
+            var connectionMethod = new ConnectionMethodIP(ipaddress, (ushort)port, m_ConnectionManager, m_ProfileManager, playerName);
+            m_ConnectionManager.ChangeState(m_ConnectionManager.m_StartingHost.Configure(connectionMethod));
         }
 
         public override void StartHostLobby(string playerName)
         {
-            SetConnectionPayload(GetPlayerId(), playerName);
-            m_ConnectionManager.ChangeState(m_ConnectionManager.m_StartingHost);
-        }
-
-        void SetConnectionPayload(string playerId, string playerName)
-        {
-            var payload = JsonUtility.ToJson(new ConnectionPayload()
-            {
-                playerId = playerId,
-                playerName = playerName,
-                isDebug = Debug.isDebugBuild
-            });
-
-            var payloadBytes = System.Text.Encoding.UTF8.GetBytes(payload);
-
-            m_ConnectionManager.NetworkManager.NetworkConfig.ConnectionData = payloadBytes;
-        }
-
-        string GetPlayerId()
-        {
-            if (UnityServices.State != ServicesInitializationState.Initialized)
-            {
-                return ClientPrefs.GetGuid() + m_ProfileManager.Profile;
-            }
-
-            return AuthenticationService.Instance.IsSignedIn ? AuthenticationService.Instance.PlayerId : ClientPrefs.GetGuid() + m_ProfileManager.Profile;
+            var connectionMethod = new ConnectionMethodRelay(m_LobbyServiceFacade, m_LocalLobby, m_ConnectionManager, m_ProfileManager, playerName);
+            m_ConnectionManager.ChangeState(m_ConnectionManager.m_StartingHost.Configure(connectionMethod));
         }
     }
 }